Preparing for 1.44 release

+++ SearchLink.popclipext/searchlink.rb
@@ -1,12 +1,15 @@
 #!/usr/bin/env ruby
+# frozen_string_literal: true
+
 require 'shellwords'
 
 input = ENV['POPCLIP_TEXT']
 
 def install_searchlink
-  unless File.exist?(File.expand_path('~/Library/Services/SearchLink.workflow'))
-    folder = File.expand_path('~/Downloads')
-    services = File.expand_path('~/Library/Services')
+  user = `whoami`.strip
+  unless File.exist?(File.expand_path("/Users/#{user}/Library/Services/SearchLink.workflow"))
+    folder = File.expand_path("/Users/#{user}/Downloads")
+    services = File.expand_path("/Users/#{user}/Library/Services")
     dl = File.join(folder, 'SearchLink.zip')
     `curl -SsL -o "#{dl}" https://github.com/ttscoff/searchlink/releases/latest/download/SearchLink.zip`
     Dir.chdir(folder)
@@ -16,9 +19,7 @@ def install_searchlink
     ['SearchLink.workflow', 'SearchLink File.workflow', 'Jump to SearchLink Error.workflow'].each do |wflow|
       src = File.join(folder, 'SearchLink Services', wflow)
       dest = File.join(services, wflow)
-      if File.exist?(src) && File.exist?(dest)
-        FileUtils.rm_rf(dest)
-      end
+      FileUtils.rm_rf(dest) if File.exist?(dest)
       FileUtils.mv(src, dest, force: true)
     end
     FileUtils.rm_rf('SearchLink Services')
@@ -26,8 +27,8 @@ def install_searchlink
 end
 
 install_searchlink
-
-res = `echo #{Shellwords.escape(input)} | automator -i - ~/Library/Services/SearchLink.workflow`.strip
+user = `whoami`.strip
+res = `echo #{Shellwords.escape(input)} | automator -i - /Users/#{user}/Library/Services/SearchLink.workflow`.strip
 
 res.gsub!(/'/, "\\\\'")
 res.gsub!(/^\(/, '[')
